refactor(sftp): centralize same-pane paste filesystem checks
